+++ Refresh.GameServer/Importing/Mip/MipHeader.cs
@@ -13,7 +13,7 @@ public class MipHeader
     public bool Alpha;
     public uint DataOffset;
 
-    public Rgba32[] Clut;
+    public Rgba32[] ColorLookupTable;
     
     public static MipHeader Read(Stream stream)
     {
@@ -29,13 +29,13 @@ public class MipHeader
             TexMode = reader.ReadByte(),
             Alpha = reader.ReadByte() == 1,
             DataOffset = reader.ReadUInt32(),
-            Clut = new Rgba32[256],
+            ColorLookupTable = new Rgba32[256],
         };
 
         stream.Seek(header.ClutOffset, SeekOrigin.Begin);
-        for (int i = 0; i < header.DataOffset && i < header.Clut.Length; i++)
+        for (int i = 0; i < header.DataOffset && i < header.ColorLookupTable.Length; i++)
         {
-            header.Clut[i] = new Rgba32(reader.ReadByte(), reader.ReadByte(), reader.ReadByte(), reader.ReadByte());
+            header.ColorLookupTable[i] = new Rgba32(reader.ReadByte(), reader.ReadByte(), reader.ReadByte(), reader.ReadByte());
         }
 
         return header;
@@ -43,13 +43,13 @@ public class MipHeader
 
     public void Write(Stream stream)
     {
-        const int clutBase = 0x30;
+        const int colorLookupTableBase = 0x30;
         const int imageBase = 0x480;
 
         long start = stream.Position;
         
         BinaryWriter writer = new(stream);
-        writer.Write((uint)clutBase);
+        writer.Write((uint)colorLookupTableBase);
         writer.Write(this.Width);
         writer.Write(this.Height);
         writer.Write(this.Bpp);
@@ -58,15 +58,15 @@ public class MipHeader
         writer.Write((byte)(this.Alpha ? 1 : 0));
         writer.Write((uint)imageBase);
 
-        long amountUntilClut = clutBase - stream.Position - start;
-        while (amountUntilClut > 0)
+        long bytesUntilColorLookupTable = colorLookupTableBase - stream.Position - start;
+        while (bytesUntilColorLookupTable > 0)
         {
             //Write a null byte to pad
             writer.Write((byte)0xAA);
-            amountUntilClut--;
+            bytesUntilColorLookupTable--;
         }
 
-        foreach (Rgba32 color in this.Clut)
+        foreach (Rgba32 color in this.ColorLookupTable)
         {
             writer.Write(color.R);
             writer.Write(color.G);
@@ -74,12 +74,12 @@ public class MipHeader
             writer.Write(color.A);
         }
         
-        long amountUntilImage = imageBase - stream.Position - start;
-        while (amountUntilImage > 0)
+        long bytesUntilImageData = imageBase - stream.Position - start;
+        while (bytesUntilImageData > 0)
         {
             //Write a null byte to pad
             writer.Write((byte)0xAA);
-            amountUntilImage--;
+            bytesUntilImageData--;
         }
     }
 }
