+++ RefreshTests.GameServer/RefreshTests.GameServer.csproj
@@ -50,11 +50,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <None Remove="Resources\1x1.png" />
       <EmbeddedResource Include="Resources\1x1.png" />
     </ItemGroup>
-
-
-
 
 </Project>
